+++ src/index.ts
@@ -3,7 +3,7 @@ export * from "./scalars";
 export * from "./errors";
 export * from "./interfaces";
 
-export { buildSchema, BuildSchemaOptions } from "./utils/buildSchema";
+export { buildSchema, buildSchemaSync, BuildSchemaOptions } from "./utils/buildSchema";
 export { useContainer } from "./utils/container";
 export { AuthChecker, ActionData, FilterActionData } from "./types";