+++ src/utils/buildSchema.ts
@@ -6,17 +6,23 @@ import { loadResolversFromGlob } from "../helpers/loadResolversFromGlob";
 export interface BuildSchemaOptions extends SchemaGeneratorOptions {
   resolvers: Array<Function | string>;
 }
-export async function buildSchema(options: BuildSchemaOptions): Promise<GraphQLSchema> {
+export function buildSchema(options: BuildSchemaOptions): Promise<GraphQLSchema> {
+  loadResolvers(options);
+  return SchemaGenerator.generateFromMetadata(options);
+}
+
+export function buildSchemaSync(options: BuildSchemaOptions): GraphQLSchema {
+  loadResolvers(options);
+  return SchemaGenerator.generateFromMetadataSync(options);
+}
+
+function loadResolvers(options: BuildSchemaOptions) {
   if (options.resolvers.length === 0) {
     throw new Error("Empty `resolvers` array property found in `buildSchema` options!");
   }
-  await Promise.all(
-    options.resolvers.map(async resolver => {
-      if (typeof resolver === "string") {
-        await loadResolversFromGlob(resolver);
-      }
-    }),
-  );
-
-  return SchemaGenerator.generateFromMetadata(options);
+  options.resolvers.forEach(resolver => {
+    if (typeof resolver === "string") {
+      loadResolversFromGlob(resolver);
+    }
+  });
 }